requirements.txt
anyascii==0.3.2
asgiref==3.8.1
bcp47==0.0.4
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
defusedxml==0.7.1
Django==5.2.1
django-classy-tags==4.1.0
django-filter==25.1
django-modelcluster==6.4
django-permissionedforms==0.1
django-sekizai==4.1.0
django-stubs-ext==5.2.0
django-taggit==6.1.0
django-tasks==0.6.1
django-treebeard==4.7.1
djangorestframework==3.16.0
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.10
laces==0.1.2
mysqlclient==2.2.7
openpyxl==3.1.5
pillow==11.2.1
pillow_heif==0.21.0
python-dotenv==1.1.0
requests==2.32.3
soupsieve==2.7
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.4.0
wagtail==6.4.2
wagtail-accessibility==2.0.0
wagtail-modeladmin==2.2.0
wagtail-seo==3.0.0
wagtailset==1.1.3
wagtailvideos==7.0.3
Willow==1.10.0
